+++ console_gpt/menus/model_menu.py
@@ -36,14 +36,21 @@ def model_menu() -> Dict[str, Union[int, str, float]]:
     # Build the menu based on the available models (chat.models.<model>)
     menu_data = list(all_models.keys())
     menu_data.append("ollama")
-    menu_title = "{} Select a model:".format(use_emoji_maybe("\U0001F916"))
+    menu_title = "{} Select a model:".format(use_emoji_maybe("\U0001f916"))
     selection = base_multiselect_menu("Model menu", menu_data, menu_title, default_model)
     if selection == "ollama":
         models = get_ollama()
         if models:
-            menu_title = "{} Select a locally hosted model:".format(use_emoji_maybe("\U0001F916"))
+            menu_title = "{} Select a locally hosted model:".format(use_emoji_maybe("\U0001f916"))
             local_selection = base_multiselect_menu("Ollama models", models, menu_title)
-            model_data = {'api_key': 'ollama', 'api_usage': 0, 'model_input_pricing_per_1k': 0, 'model_max_tokens': 0, 'model_name': local_selection, 'model_output_pricing_per_1k': 0}
+            model_data = {
+                "api_key": "ollama",
+                "api_usage": 0,
+                "model_input_pricing_per_1k": 0,
+                "model_max_tokens": 0,
+                "model_name": local_selection,
+                "model_output_pricing_per_1k": 0,
+            }
         else:
             custom_print("info", "No models found in Ollama. Please check the Ollama server or select another model.")
             model_menu()