+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 > Features:
+- [c41655a] [Feature] Run Ollama hosted models locally
 - [3b9d331] [Feature] Add Deepseek R1
 - [3aa9667] [Feature] Add the ability to read old chats via 'chats' command